Refactor Summary components, fix color on filter button
