+++ src/pages/OffersPage/components/ActiveTabContent/components/LoansTable/LoansTable.module.less
@@ -63,6 +63,15 @@
   padding-left: 18px;
 }
 
+.filterButton {
+  svg path:nth-child(2) {
+    fill: var(--additional-yellow-primary-deep);
+  }
+  circle {
+    fill: var(--content-primary);
+  }
+}
+
 //? Active row styles
 .terminated {
   background: var(--additional-lava-secondary) !important;
