+++ src/pages/OffersPage/components/ActiveTabContent/components/LoansTable/Summary/Summary.tsx
@@ -1,78 +1,34 @@
 import { FC, useMemo } from 'react'
 
-import { useConnection, useWallet } from '@solana/wallet-adapter-react'
-import { sumBy } from 'lodash'
-import { TxnExecutor } from 'solana-transactions-executor'
-
-import { Button } from '@banx/components/Buttons'
-import { StatInfo } from '@banx/components/StatInfo'
-
 import { Loan } from '@banx/api/core'
-import { TABLET_WIDTH } from '@banx/constants'
-import { useWindowSize } from '@banx/hooks'
-import { defaultTxnErrorHandler } from '@banx/transactions'
-import { makeClaimAction } from '@banx/transactions/loans'
-import { enqueueSnackbar } from '@banx/utils'
+import { isLoanAbleToClaim } from '@banx/pages/OffersPage'
+import { isUnderWaterLoan } from '@banx/utils'
 
-import { TerminateContent } from './components'
+import { ClaimContent, TerminateContent } from './components'
 
 import styles from './Summary.module.less'
 
 interface SummaryProps {
   updateOrAddLoan: (loan: Loan) => void
   hideLoans: (...mints: string[]) => void
-  loansToClaim: Loan[]
-  loansToTerminate: Loan[]
   isUnderwaterFilterActive: boolean
 
   selectedLoans: Loan[]
   setSelection: (loans: Loan[]) => void
+
+  loans: Loan[]
 }
 
 export const Summary: FC<SummaryProps> = ({
   updateOrAddLoan,
-  loansToTerminate,
-  loansToClaim,
+  loans,
   hideLoans,
   isUnderwaterFilterActive,
-
   selectedLoans,
   setSelection,
 }) => {
-  const wallet = useWallet()
-  const { connection } = useConnection()
-  const { width } = useWindowSize()
-  const isSmallDesktop = width < TABLET_WIDTH
-
-  const totalClaimableFloor = useMemo(
-    () => sumBy(loansToClaim, ({ nft }) => nft.collectionFloor),
-    [loansToClaim],
-  )
-
-  const claimLoans = () => {
-    const txnParams = loansToClaim.map((loan) => ({ loan }))
-
-    new TxnExecutor(makeClaimAction, { wallet, connection })
-      .addTxnParams(txnParams)
-      .on('pfSuccessEach', (results) => {
-        enqueueSnackbar({
-          message: 'Collateral successfully claimed',
-          type: 'success',
-          solanaExplorerPath: `tx/${results[0].txnHash}`,
-        })
-      })
-      .on('pfSuccessAll', () => {
-        hideLoans(...loansToClaim.map(({ nft }) => nft.mint))
-      })
-      .on('pfError', (error) => {
-        defaultTxnErrorHandler(error, {
-          additionalData: txnParams,
-          walletPubkey: wallet?.publicKey?.toBase58(),
-          transactionName: 'Claim',
-        })
-      })
-      .execute()
-  }
+  const loansToClaim = useMemo(() => loans.filter(isLoanAbleToClaim), [loans])
+  const loansToTerminate = useMemo(() => loans.filter(isUnderWaterLoan), [loans])
 
   return (
     <div className={styles.summaryContainer}>
@@ -84,50 +40,8 @@ export const Summary: FC<SummaryProps> = ({
           updateOrAddLoan={updateOrAddLoan}
         />
       ) : (
-        <ClaimNFTsButton
-          onClick={claimLoans}
-          totalLoans={loansToClaim.length}
-          value={totalClaimableFloor}
-          isSmallDesktop={isSmallDesktop}
-        />
+        <ClaimContent loans={loansToClaim} hideLoans={hideLoans} />
       )}
-    </div>
-  )
-}
-
-interface ButtonProps {
-  onClick: () => void
-  totalLoans?: number
-  isSmallDesktop: boolean
-  value: number
-}
-
-const ClaimNFTsButton: FC<ButtonProps> = (props) => {
-  const { isSmallDesktop, totalLoans, onClick, value } = props
-  const buttonText = isSmallDesktop ? 'Claim' : 'Claim all NFTs'
-  const label = isSmallDesktop ? 'Claimable floor' : 'Collateral'
-
-  return (
-    <div className={styles.infoRow}>
-      <div className={styles.loansContainer}>
-        <p className={styles.loansValueText}>{totalLoans}</p>
-        <div className={styles.loansInfoContainer}>
-          <StatInfo
-            label={label}
-            value={value}
-            classNamesProps={{ value: styles.value }}
-            divider={1e9}
-          />
-        </div>
-      </div>
-      <Button
-        className={styles.summaryButton}
-        onClick={onClick}
-        disabled={!totalLoans}
-        variant="secondary"
-      >
-        {buttonText}
-      </Button>
     </div>
   )
 }
