+++ src/pages/OffersPage/components/ActiveTabContent/components/LoansTable/Summary/components.tsx
@@ -1,7 +1,8 @@
-import { FC } from 'react'
+import { FC, useMemo } from 'react'
 
 import { useConnection, useWallet } from '@solana/wallet-adapter-react'
 import classNames from 'classnames'
+import { sumBy } from 'lodash'
 import { TxnExecutor } from 'solana-transactions-executor'
 
 import { Button } from '@banx/components/Buttons'
@@ -9,8 +10,10 @@ import { CounterSlider } from '@banx/components/Slider'
 import { StatInfo, VALUES_TYPES } from '@banx/components/StatInfo'
 
 import { Loan } from '@banx/api/core'
+import { TABLET_WIDTH } from '@banx/constants'
+import { useWindowSize } from '@banx/hooks'
 import { defaultTxnErrorHandler } from '@banx/transactions'
-import { makeTerminateAction } from '@banx/transactions/loans'
+import { makeClaimAction, makeTerminateAction } from '@banx/transactions/loans'
 import { HealthColorIncreasing, enqueueSnackbar, getColorByPercent } from '@banx/utils'
 
 import { getTerminateStatsInfo } from './helpers'
@@ -105,6 +108,72 @@ export const TerminateContent: FC<TerminateContentProps> = ({
           Terminate
         </Button>
       </div>
+    </div>
+  )
+}
+
+interface ClaimContentProps {
+  loans: Loan[]
+  hideLoans: (...mints: string[]) => void
+}
+
+export const ClaimContent: FC<ClaimContentProps> = ({ loans, hideLoans }) => {
+  const wallet = useWallet()
+  const { connection } = useConnection()
+  const { width } = useWindowSize()
+  const isSmallDesktop = width < TABLET_WIDTH
+
+  const totalClaimableFloor = useMemo(() => sumBy(loans, ({ nft }) => nft.collectionFloor), [loans])
+
+  const buttonText = isSmallDesktop ? 'Claim' : 'Claim all NFTs'
+  const label = isSmallDesktop ? 'Claimable floor' : 'Collateral'
+
+  const claimLoans = () => {
+    const txnParams = loans.map((loan) => ({ loan }))
+
+    new TxnExecutor(makeClaimAction, { wallet, connection })
+      .addTxnParams(txnParams)
+      .on('pfSuccessEach', (results) => {
+        enqueueSnackbar({
+          message: 'Collateral successfully claimed',
+          type: 'success',
+          solanaExplorerPath: `tx/${results[0].txnHash}`,
+        })
+      })
+      .on('pfSuccessAll', () => {
+        hideLoans(...loans.map(({ nft }) => nft.mint))
+      })
+      .on('pfError', (error) => {
+        defaultTxnErrorHandler(error, {
+          additionalData: txnParams,
+          walletPubkey: wallet?.publicKey?.toBase58(),
+          transactionName: 'Claim',
+        })
+      })
+      .execute()
+  }
+
+  return (
+    <div className={styles.infoRow}>
+      <div className={styles.loansContainer}>
+        <p className={styles.loansValueText}>{loans.length}</p>
+        <div className={styles.loansInfoContainer}>
+          <StatInfo
+            label={label}
+            value={totalClaimableFloor}
+            classNamesProps={{ value: styles.value }}
+            divider={1e9}
+          />
+        </div>
+      </div>
+      <Button
+        className={styles.summaryButton}
+        onClick={claimLoans}
+        disabled={!loans.length}
+        variant="secondary"
+      >
+        {buttonText}
+      </Button>
     </div>
   )
 }
